requirements.txt
# Core ML dependencies
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# MediaPipe for landmark extraction
mediapipe>=0.10.0

# Computer vision
opencv-python>=4.8.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# API serving
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# Configuration
pyyaml>=6.0
pydantic>=2.0.0

# Progress bars
tqdm>=4.65.0
